feat:relations on stream table and seeders

+++ database/migrations/2023_12_07_142542_create_country_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('countries', function (Blueprint $table) {
             $table->id();
             $table->string('country_name', 50);
-            $table->integer('country_bonus');
+            $table->decimal('country_bonus',5,2);
         });
     }
